rename : Manager 하위 reservation 관련 코드 -> reservation으로 이동
