+++ common/src/main/java/com/kernel/common/reservation/entity/Reservation.java
@@ -3,25 +3,16 @@ package com.kernel.common.reservation.entity;
 import com.kernel.common.customer.entity.Customer;
 import com.kernel.common.global.entity.BaseEntity;
 import com.kernel.common.global.entity.ServiceCategory;
-import com.kernel.common.global.enums.ReservationStatus;
 import com.kernel.common.manager.entity.Manager;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import java.time.LocalDate;
-import java.time.LocalTime;
+import com.kernel.common.reservation.enums.ReservationStatus;
+import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
 
 @Entity
 @Table(name = "reservation")
